# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Telegram Bots Directory
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Bot {
  id: number;
  username: string;
  name: string;
  description: string;
  category: string;
  categoryId: number;
  rating: number;
  reviewCount: number;
  isVerified: boolean;
  isPremium: boolean;
  tags: string[];
  monthlyUsers: number;
  iconEmoji: string;
  telegramUrl: string;
}

export interface BotInput {
  username: string;
  name: string;
  description: string;
  categoryId: number;
  rating: number;
  reviewCount: number;
  isVerified: boolean;
  isPremium: boolean;
  tags: string[];
  monthlyUsers: number;
  iconEmoji: string;
  telegramUrl: string;
}

export type BotStatDailyViewsItem = {
  date: string;
  views: number;
};

export type BotStatGeoBreakdownItem = {
  country: string;
  countryCode: string;
  views: number;
};

export interface BotStat {
  botId: number;
  botName: string;
  botEmoji: string;
  totalViews: number;
  uniqueVisitors: number;
  last7Days: number;
  last30Days: number;
  dailyViews: BotStatDailyViewsItem[];
  geoBreakdown: BotStatGeoBreakdownItem[];
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  emoji: string;
  count: number;
}

export interface AdminLoginRequest {
  password: string;
}

export interface AdminLoginResponse {
  token: string;
  message: string;
}

export interface DeleteResponse {
  success: boolean;
  message: string;
}

export interface Error {
  message: string;
}

export type ListBotsParams = {
  category?: string;
  search?: string;
  sortBy?: ListBotsSortBy;
};

export type ListBotsSortBy =
  (typeof ListBotsSortBy)[keyof typeof ListBotsSortBy];

export const ListBotsSortBy = {
  rating: "rating",
  name: "name",
  popular: "popular",
} as const;