# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Telegram Bots Directory
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all bots
 */
export const ListBotsQueryParams = zod.object({
  category: zod.coerce.string().optional(),
  search: zod.coerce.string().optional(),
  sortBy: zod.enum(["rating", "name", "popular"]).optional(),
});

export const ListBotsResponseItem = zod.object({
  id: zod.number(),
  username: zod.string(),
  name: zod.string(),
  description: zod.string(),
  category: zod.string(),
  categoryId: zod.number(),
  rating: zod.number(),
  reviewCount: zod.number(),
  isVerified: zod.boolean(),
  isPremium: zod.boolean(),
  tags: zod.array(zod.string()),
  monthlyUsers: zod.number(),
  iconEmoji: zod.string(),
  telegramUrl: zod.string(),
});
export const ListBotsResponse = zod.array(ListBotsResponseItem);

/**
 * @summary Get bot by ID
 */
export const GetBotParams = zod.object({
  id: zod.coerce.number(),
});

export const GetBotResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  name: zod.string(),
  description: zod.string(),
  category: zod.string(),
  categoryId: zod.number(),
  rating: zod.number(),
  reviewCount: zod.number(),
  isVerified: zod.boolean(),
  isPremium: zod.boolean(),
  tags: zod.array(zod.string()),
  monthlyUsers: zod.number(),
  iconEmoji: zod.string(),
  telegramUrl: zod.string(),
});

/**
 * @summary Record a bot page view
 */
export const RecordBotViewParams = zod.object({
  id: zod.coerce.number(),
});

export const RecordBotViewResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all categories
 */
export const ListCategoriesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  slug: zod.string(),
  emoji: zod.string(),
  count: zod.number(),
});
export const ListCategoriesResponse = zod.array(ListCategoriesResponseItem);

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  token: zod.string(),
  message: zod.string(),
});

/**
 * @summary Create a new bot
 */
export const CreateBotBody = zod.object({
  username: zod.string(),
  name: zod.string(),
  description: zod.string(),
  categoryId: zod.number(),
  rating: zod.number(),
  reviewCount: zod.number(),
  isVerified: zod.boolean(),
  isPremium: zod.boolean(),
  tags: zod.array(zod.string()),
  monthlyUsers: zod.number(),
  iconEmoji: zod.string(),
  telegramUrl: zod.string(),
});

/**
 * @summary Update a bot
 */
export const UpdateBotParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateBotBody = zod.object({
  username: zod.string(),
  name: zod.string(),
  description: zod.string(),
  categoryId: zod.number(),
  rating: zod.number(),
  reviewCount: zod.number(),
  isVerified: zod.boolean(),
  isPremium: zod.boolean(),
  tags: zod.array(zod.string()),
  monthlyUsers: zod.number(),
  iconEmoji: zod.string(),
  telegramUrl: zod.string(),
});

export const UpdateBotResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  name: zod.string(),
  description: zod.string(),
  category: zod.string(),
  categoryId: zod.number(),
  rating: zod.number(),
  reviewCount: zod.number(),
  isVerified: zod.boolean(),
  isPremium: zod.boolean(),
  tags: zod.array(zod.string()),
  monthlyUsers: zod.number(),
  iconEmoji: zod.string(),
  telegramUrl: zod.string(),
});

/**
 * @summary Delete a bot
 */
export const DeleteBotParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteBotResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Get visit statistics for all bots
 */
export const GetBotStatsResponseItem = zod.object({
  botId: zod.number(),
  botName: zod.string(),
  botEmoji: zod.string(),
  totalViews: zod.number(),
  uniqueVisitors: zod.number(),
  last7Days: zod.number(),
  last30Days: zod.number(),
  dailyViews: zod.array(
    zod.object({
      date: zod.string(),
      views: zod.number(),
    }),
  ),
  geoBreakdown: zod.array(
    zod.object({
      country: zod.string(),
      countryCode: zod.string(),
      views: zod.number(),
    }),
  ),
});
export const GetBotStatsResponse = zod.array(GetBotStatsResponseItem);